added stack float and separated putils
