+++ examples/autest_in/NA_stack_float/pp_global.sym
@@ -1,0 +1,175 @@
+1	Global	BrickBusVoltage	8192	1
+2	Global	PeakMaxAmps	8193	1
+3	Global	ProtWrongLimit()	8194	9
+4	Global	ProtConfig()	8203	9
+5	Global	prevLimitSwLow()	8212	9
+6	Global	prevLimitSwHigh()	8221	9
+7	Global	captureEnableMtr()	8230	9
+8	Global	glb_redis_comms	8239	1
+9	Global	timeout_period()	8240	9
+10	Global	idle_strategy_en()	8249	9
+11	Global	hold_percentage()	8258	9
+12	Global	full_current()	8267	9
+13	Global	hold_current()	8276	9
+14	Global	time_counter()	8285	9
+15	Global	velocity_precision	8294	1
+16	Global	JogSpeedLockFault	8295	1
+17	Global	MaxSpeedLockFault	8296	1
+18	Global	MaxSpeedBak()	8297	9
+19	Global	PosSfLockFault	8306	1
+20	Global	posSfBak()	8307	9
+21	Global	pLimitsLockFault	8316	1
+22	Global	pLimitBak()	8317	9
+23	Ptr	IN1_1	8192	1
+24	Ptr	IN1_2	8193	1
+25	Ptr	IN1_3	8194	1
+26	Ptr	IN1_4	8195	1
+27	Ptr	IN1_5	8196	1
+28	Ptr	IN1_6	8197	1
+29	Ptr	IN1_7	8198	1
+30	Ptr	IN1_8	8199	1
+31	Ptr	IN1_9	8200	1
+32	Ptr	IN1_10	8201	1
+33	Ptr	IN1_11	8202	1
+34	Ptr	IN1_12	8203	1
+35	Ptr	IN1_13	8204	1
+36	Ptr	IN1_14	8205	1
+37	Ptr	IN1_15	8206	1
+38	Ptr	IN1_16	8207	1
+39	Ptr	OUT1_1	8208	1
+40	Ptr	OUT1_2	8209	1
+41	Ptr	OUT1_3	8210	1
+42	Ptr	OUT1_4	8211	1
+43	Ptr	OUT1_5	8212	1
+44	Ptr	OUT1_6	8213	1
+45	Ptr	OUT1_7	8214	1
+46	Ptr	OUT1_8	8215	1
+47	Ptr	IN2_1	8216	1
+48	Ptr	IN2_2	8217	1
+49	Ptr	IN2_3	8218	1
+50	Ptr	IN2_4	8219	1
+51	Ptr	IN2_5	8220	1
+52	Ptr	IN2_6	8221	1
+53	Ptr	IN2_7	8222	1
+54	Ptr	IN2_8	8223	1
+55	Ptr	IN2_9	8224	1
+56	Ptr	IN2_10	8225	1
+57	Ptr	IN2_11	8226	1
+58	Ptr	IN2_12	8227	1
+59	Ptr	IN2_13	8228	1
+60	Ptr	IN2_14	8229	1
+61	Ptr	IN2_15	8230	1
+62	Ptr	IN2_16	8231	1
+63	Ptr	OUT2_1	8232	1
+64	Ptr	OUT2_2	8233	1
+65	Ptr	OUT2_3	8234	1
+66	Ptr	OUT2_4	8235	1
+67	Ptr	OUT2_5	8236	1
+68	Ptr	OUT2_6	8237	1
+69	Ptr	OUT2_7	8238	1
+70	Ptr	OUT2_8	8239	1
+71	Ptr	GPRelay1X9	8240	1
+72	Ptr	GPRelay2X9	8241	1
+73	Ptr	GPRelay1X10	8242	1
+74	Ptr	GPRelay2X10	8243	1
+75	Ptr	GPRelay1X11	8244	1
+76	Ptr	GPRelay2X11	8245	1
+77	Ptr	GPRelay1X12	8246	1
+78	Ptr	GPRelay2X12	8247	1
+79	Ptr	GpIn1X9	8248	1
+80	Ptr	GpIn2X9	8249	1
+81	Ptr	GpIn1X10	8250	1
+82	Ptr	GpIn2X10	8251	1
+83	Ptr	GpIn1X11	8252	1
+84	Ptr	GpIn2X11	8253	1
+85	Ptr	GpIn1X12	8254	1
+86	Ptr	GpIn2X12	8255	1
+87	Ptr	ADC1X9	8256	1
+88	Ptr	ADC2X9	8257	1
+89	Ptr	ADC1X10	8258	1
+90	Ptr	ADC2X10	8259	1
+91	Ptr	ADC1X11	8260	1
+92	Ptr	ADC2X11	8261	1
+93	Ptr	ADC1X12	8262	1
+94	Ptr	ADC2X12	8263	1
+95	Global	CalcMaxDac1	8326	1
+96	Global	RmsPeakCur1	8327	1
+97	Global	TimeAtPeak1	8328	1
+98	Global	RmsContCur1	8329	1
+99	Global	maxVolts1	8330	1
+100	Global	EncRes1	8331	1
+101	Global	enc_cnt_per_rev1	8332	1
+102	Global	MtrPolePairs1	8333	1
+103	Global	encoder_direction1	8334	1
+104	Global	CalcMaxDac2	8335	1
+105	Global	RmsPeakCur2	8336	1
+106	Global	TimeAtPeak2	8337	1
+107	Global	RmsContCur2	8338	1
+108	Global	maxVolts2	8339	1
+109	Global	EncRes2	8340	1
+110	Global	MaxRPM2	8341	1
+111	Global	StepAngle2	8342	1
+112	Global	CalcMaxDac3	8343	1
+113	Global	RmsPeakCur3	8344	1
+114	Global	TimeAtPeak3	8345	1
+115	Global	RmsContCur3	8346	1
+116	Global	maxVolts3	8347	1
+117	Global	EncRes3	8348	1
+118	Global	MaxRPM3	8349	1
+119	Global	StepAngle3	8350	1
+120	Global	CalcMaxDac4	8351	1
+121	Global	RmsPeakCur4	8352	1
+122	Global	TimeAtPeak4	8353	1
+123	Global	RmsContCur4	8354	1
+124	Global	maxVolts4	8355	1
+125	Global	EncRes4	8356	1
+126	Global	MaxRPM4	8357	1
+127	Global	StepAngle4	8358	1
+128	Global	CalcMaxDac5	8359	1
+129	Global	RmsPeakCur5	8360	1
+130	Global	TimeAtPeak5	8361	1
+131	Global	RmsContCur5	8362	1
+132	Global	maxVolts5	8363	1
+133	Global	EncRes5	8364	1
+134	Global	MaxRPM5	8365	1
+135	Global	StepAngle5	8366	1
+136	Global	CalcMaxDac6	8367	1
+137	Global	RmsPeakCur6	8368	1
+138	Global	TimeAtPeak6	8369	1
+139	Global	RmsContCur6	8370	1
+140	Global	maxVolts6	8371	1
+141	Global	EncRes6	8372	1
+142	Global	MaxRPM6	8373	1
+143	Global	StepAngle6	8374	1
+144	Global	CalcMaxDac7	8375	1
+145	Global	RmsPeakCur7	8376	1
+146	Global	TimeAtPeak7	8377	1
+147	Global	RmsContCur7	8378	1
+148	Global	maxVolts7	8379	1
+149	Global	EncRes7	8380	1
+150	Global	MaxRPM7	8381	1
+151	Global	StepAngle7	8382	1
+152	Global	CalcMaxDac8	8383	1
+153	Global	RmsPeakCur8	8384	1
+154	Global	TimeAtPeak8	8385	1
+155	Global	RmsContCur8	8386	1
+156	Global	maxVolts8	8387	1
+157	Global	EncRes8	8388	1
+158	Global	MaxRPM8	8389	1
+159	Global	StepAngle8	8390	1
+160	Global	EncRes9	8391	1
+161	Global	encoder_direction9	8392	1
+162	Global	EncRes10	8393	1
+163	Global	encoder_direction10	8394	1
+164	Global	EncRes11	8395	1
+165	Global	encoder_direction11	8396	1
+166	Global	EncRes12	8397	1
+167	Global	encoder_direction12	8398	1
+168	Global	EncRes13	8399	1
+169	Global	encoder_direction13	8400	1
+170	Global	EncRes14	8401	1
+171	Global	encoder_direction14	8402	1
+172	Global	EncRes15	8403	1
+173	Global	encoder_direction15	8404	1
+174	Global	EncRes16	8405	1
+175	Global	encoder_direction16	8406	1
