+++ examples/ol_direct.py
@@ -15,19 +15,21 @@ from time import sleep
 if __name__ == "__main__":
 
     tst = utils.undump_obj(
-        "ol_direct_sample",
-        path.join(path.dirname(path.abspath(__file__)), "autest_in"),
+        "ol_stack_float", path.join(path.dirname(path.abspath(__file__)), "autest_in"),
     )
-    tst["ppmac_hostname"] = "10.23.220.232"
-    tst["ppmac_is_backward"] = True
+    # tst["ppmac_hostname"] = "10.23.220.232"
+    # tst["ppmac_is_backward"] = True
 
     ol_test = OL_Rdb_Lim2Lim(_VERBOSE_=2, tst=tst, motor_id="Mot_A")
-    agents = ppra.ra.compile_n_install({}, globals().copy(), "WORKSHOP01")
+    agents = ppra.ra.compile_n_install({}, globals().copy(), "blah")
     max_loop = 100
 
-    ppra.do_ags([ol_test.set_wpKey_ag, ol_test.system_config_ag], cycle_period=1)
-    ppra.do_ags(ol_test.ma_base_config_ag)
-    ppra.do_ags(ol_test.set_initial_setup_ag)
+    if tst["sysconfig_download"]:
+        ppra.do_ags([ol_test.set_wpKey_ag, ol_test.system_config_ag], cycle_period=1)
+    if tst["baseconfig_download"]:
+        ppra.do_ags(ol_test.ma_base_config_ag)
+
+    # ppra.do_ags(ol_test.set_initial_setup_ag) # loads plc10 and prog 10 !
 
     ppra.do_ags(ol_test.ma_test_config_ag)
     ppra.do_ags(ol_test.ma_go_mlim_ag)
